eanworker/requirements.txt
aioredis==1.2.0
python-barcode==0.10.0
Pillow==6.0.0
motor==2.0.0